changed HealthScaleIncrementFactor

+++ Assets/Scripts/Survivors/App/Config/ConstantsConfig.cs
@@ -5,7 +5,7 @@ namespace Survivors.App.Config
     public class ConstantsConfig
     {
         [DataMember(Name = "HealthScaleIncrementFactor")]
-        private int _healthScaleIncrementFactor;
+        private float _healthScaleIncrementFactor;
 
         [DataMember(Name = "FirstUnit")]
         private string _firstUnit;    
@@ -16,7 +16,7 @@ namespace Survivors.App.Config
         [DataMember(Name = "ReviveEnemyRemoveRadius")]
         private float _reviveEnemyRemoveRadius;
 
-        public int HealthScaleIncrementFactor => _healthScaleIncrementFactor;
+        public float HealthScaleIncrementFactor => _healthScaleIncrementFactor;
         public string FirstUnit => _firstUnit;
         public int MaxMetaUpgradeLevel => _maxMetaUpgradeLevel;
